refactor: extract logic into composable
